split responsebility: data, trininer, model selection
